% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc}
\alias{mcmc}
\alias{run_mcmc}
\title{Main function for the MCMC loop}
\usage{
run_mcmc(
  nIter,
  burnin,
  thin,
  n,
  nsamp,
  ninit,
  metropolis,
  simple,
  convex,
  npoint,
  dirichlet,
  proportion_model,
  BVS,
  gamma_prior,
  gamma_sampler,
  eta_prior,
  eta_sampler,
  initList,
  rangeList,
  hyperparList,
  datEvent,
  datTime,
  datX,
  datX0,
  datProportionConst
)
}
\arguments{
\item{nIter}{Number of MCMC iterations}

\item{burnin}{Length of MCMC burn-in period}

\item{thin}{Number of thinning}

\item{n}{Number of samples to draw}

\item{nsamp}{How many samples to draw for generating each sample; only the last draw will be kept}

\item{ninit}{Number of initials as meshgrid values for envelop search}

\item{metropolis}{TBA}

\item{simple}{TBA}

\item{convex}{Adjustment for convexity (non-negative value, default 1.0)}

\item{npoint}{Maximum number of envelope points}

\item{dirichlet}{Not yet implemented}

\item{proportion_model}{TBA}

\item{BVS}{TBA}

\item{gamma_prior}{TBA}

\item{gamma_sampler}{TBA}

\item{eta_prior}{TBA}

\item{eta_sampler}{TBA}

\item{initList}{TBA}

\item{rangeList}{TBA}

\item{hyperparList}{TBA}

\item{datEvent}{TBA}

\item{datTime}{TBA}

\item{datX}{TBA}

\item{datX0}{TBA}

\item{datProportionConst}{TBA}
}
\description{
Main function for the MCMC loop
}
