% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target.R
\name{target}
\alias{target}
\title{Target density}
\usage{
target(x, theta, proportion, mu, kappas)
}
\arguments{
\item{x}{value generated from the proposal distribution}

\item{theta}{cure rate parameter (log scale)}

\item{proportion}{proportions data}

\item{mu}{mean survival time}

\item{kappas}{Weibull's true shape parameter}
}
\value{
value of the targeted (improper) probability density function
}
\description{
Predefined target density corresponding to the population survival function
of GPTCM
}
\examples{

time1 <- target(1.2, 0.1, c(0.2, 0.3, 0.5), c(0.2, 0.1, 0.4), 2)

}
