% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPTCM-package.R
\docType{package}
\name{GPTCM-package}
\alias{GPTCM-package}
\title{GPTCM: Generalized Promotion Time Cure Model with Bayesian Shrinkage Priors}
\description{
Generalized promotion time cure model with Bayesian hierarchical modeling for multiscale data integration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ocbe-uio/GPTCM}
  \item Report bugs at \url{https://github.com/ocbe-uio/GPTCM/issues}
}

}
\author{
\strong{Maintainer}: Zhi Zhao \email{zhi.zhao@medisin.uio.no}

}
\keyword{internal}
